+++ src/frontend/components/tables/filmtable.tsx
